JAT-0048 Move Echo tests inside a module
