+++ liferay/echo_team/echo_test_map.py
@@ -25,10 +25,10 @@ ECHO_TESTMAP_SHEET_NAME = 'Test Map'
 GOOGLE_SHEET_URL = 'https://docs.google.com/spreadsheets/d/'
 JIRA_TEST_MAP_TAB = "JIRA-TestMap"
 JIRA_TEST_MAP_TAB_RANGE = JIRA_TEST_MAP_TAB + '!B3:H'
-OUTPUT_BUG_THRESHOLD_EXCEED_FILE_NAME = "bug_threshold_exceed_message_echo.txt"
-OUTPUT_BUG_THRESHOLD_WARNING_FILE_NAME = "bug_threshold_warning_message_echo.txt"
-OUTPUT_INFO_FILE_NAME = "output_info_echo.txt"
-OUTPUT_MESSAGE_FILE_NAME = "output_message_echo.txt"
+OUTPUT_BUG_THRESHOLD_EXCEED_FILE_NAME = "../bug_threshold_exceed_message_echo.txt"
+OUTPUT_BUG_THRESHOLD_WARNING_FILE_NAME = "../bug_threshold_warning_message_echo.txt"
+OUTPUT_INFO_FILE_NAME = "../output_info_echo.txt"
+OUTPUT_MESSAGE_FILE_NAME = "../output_message_echo.txt"
 TESTMAP_MAPPED_RANGE = ECHO_TESTMAP_SHEET_NAME + '!G4:G'
